Fix OpenDisplay sensor data parsing to support new 14-byte format

+++ custom_components/opendisplay/ble/protocol_base.py
@@ -13,7 +13,7 @@ class AdvertisingData:
 
     battery_mv: int
     battery_pct: int
-    temperature: int | None
+    temperature: float | None
     hw_type: int
     fw_version: int
     version: int  # Config/protocol version
